finished demands and applications controllers

+++ app/Http/Controllers/DemandController.php
@@ -3,9 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Demand;
+use App\Models\Internship;
 use App\Models\User;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Date;
 use Illuminate\Support\Str;
 use Illuminate\Validation\ValidationException;
 
@@ -34,13 +34,13 @@ class DemandController extends Controller
     public function store(Request $request): \Illuminate\Http\JsonResponse
     {
 
-//        $this->validate($request, [
-//            'start_date' => 'required|date|after_or_equal:today',
-//            'end_date' => 'required|date',
-//            'duration' => 'required|integer',
-//            'supervisor_email' => 'required|email',
-//            'title' => 'required|min:12|max:200',
-//        ]);
+        $this->validate($request, [
+            'start_date' => 'required|date|after_or_equal:today',
+            'end_date' => 'required|date',
+            'duration' => 'required|integer',
+            'supervisor_email' => 'required|email',
+            'title' => 'required|min:10|max:200',
+        ]);
 
         $demand = Demand::create([
             'student_id' => $request->student_id,
@@ -98,8 +98,30 @@ class DemandController extends Controller
     public function update(Request $request, Demand $internshipDemand): \Illuminate\Http\JsonResponse
     {
         $demand = Demand::find($internshipDemand->id);
+
+        // status:2,4 => rejected
+        if ($request->status == 2 or $request->status == 4) {
+            $demand->rejection_motive = $request->rejection_motive;
+            $demand->status = $request->status;
+            $demand->save();
+            return response()->json(['demand' => $demand]);
+        }
+
         $demand->status = $request->status;
         $demand->save();
+
+        // status:3 => accepted by HOD and supervisor
+        if ($request->status == 3) {
+            $internship = Internship::create([
+                'student_id' => $request->student_id,
+                'supervisor_id' => $request->supervisor_id,
+                'start_date' => $internshipDemand->start_date,
+                'end_date' => $internshipDemand->end_date,
+                'duration' => $internshipDemand->duration,
+                'title' => $internshipDemand->title
+            ]);
+            return response()->json(['demand' => $demand, 'internship' => $internship]);
+        }
 
         return response()->json(['demand' => $demand]);
     }
@@ -109,9 +131,13 @@ class DemandController extends Controller
      */
     public function destroy(Demand $internshipDemand): \Illuminate\Http\JsonResponse
     {
+        if (!$internshipDemand->status == 0 or !$internshipDemand->status == 2 or !$internshipDemand->status == 4) {
+            return response()->json(['message' => "Internship demand has already been accepted, you can't remove it"]);
+        }
+
         $demand = Demand::find($internshipDemand->id);
         $demand->delete();
 
-        return response()->json("Demand deleted");
+        return response()->json(['message' => "Demand deleted"]);
     }
 }
